# CodeVimarsh-Backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Code Vimarsh API
  version: 1.0.0
  description: Backend API documentation for Code Vimarsh – MSUB Coding Club
servers:
  - url: http://localhost:5000/api/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
paths:
  /admin/users:
    get:
      summary: List all users with roles (CONTENT_ADMIN and above)
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of all users
        "403":
          description: Forbidden
  /admin/users/{userId}:
    get:
      summary: View a user's full profile (CONTENT_ADMIN and above)
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full user profile with admin data
        "404":
          description: User not found
  /admin/users/{userId}/role:
    patch:
      summary: Change a user's role (SUPER_ADMIN only)
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum:
                    - USER
                    - CONTENT_ADMIN
                    - SUPER_ADMIN
      responses:
        "200":
          description: Role updated successfully
        "403":
          description: Forbidden – SUPER_ADMIN only
...
  /contact:
    post:
      summary: Submit contact form
      description: Allows users to send a contact message to the system.
      tags:
        - Contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - message
              properties:
                name:
                  type: string
                  example: John Doe
                email:
                  type: string
                  format: email
                  example: johndoe@example.com
                subject:
                  type: string
                  example: Inquiry about events
                message:
                  type: string
                  example: I would like to know more about your upcoming events.
      responses:
        "201":
          description: Contact form submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Contact form submitted successfully
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: John Doe
                      email:
                        type: string
                        example: johndoe@example.com
                      message:
                        type: string
                        example: I would like to know more about your upcoming events.
                      subject:
                        type: string
                        example: Inquiry about events
        "400":
          description: Missing required fields
          content:
            application/json:
              example:
                success: false
                message: Name, email, and message are required
        "500":
          description: Internal server error
  /events:
    get:
      summary: Get all events
      tags:
        - Events
      responses:
        "200":
          description: List of all events
    post:
      summary: Create a new event
      tags:
        - Events
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvent"
      responses:
        "201":
          description: Event created
...
  /projects:
    post:
      summary: Create a new project
      tags:
        - Projects
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - tech_stack
              properties:
                title:
                  type: string
                  example: AI Resume Analyzer
                short_description:
                  type: string
                  example: AI powered resume screening tool
                category:
                  type: string
                  example: AI
                about:
                  type: string
                  example: This project analyzes resumes using NLP techniques.
                key_features:
                  type: array
                  items:
                    type: string
                  example:
                    - Resume Parsing
                    - Skill Extraction
                tech_stack:
                  type: array
                  items:
                    type: string
                  example:
                    - Node.js
                    - React
                    - OpenAI
                github_link:
                  type: string
                  example: https://github.com/user/project
                image:
                  type: string
                  example: https://image-url.com/project.png
                author_name:
                  type: string
                  example: Dhruv Pathak
                members:
                  type: array
                  items:
                    type: string
                  example:
                    - user_id_1
                    - user_id_2
      responses:
        "201":
          description: Project created successfully
    get:
      summary: Get all approved projects
      tags:
        - Projects
      responses:
        "200":
          description: List of projects retrieved successfully
...
  /users/{id}:
    get:
      summary: Get a user's public profile
      tags:
        - Users
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
        "404":
          description: User not found
    put:
      summary: Update a user's profile (owner or SUPER_ADMIN only)
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                avatar:
                  type: string
      responses:
        "200":
          description: Profile updated
        "403":
          description: Forbidden
...