# CodeVimarsh-Backend/src/app.js
import "dotenv/config";
import express from "express";
import helmet from "helmet";
import cors from "cors";
import swaggerUi from "swagger-ui-express";

import swaggerSpec from "./config/swagger.js";
import { apiLimiter } from "./middleware/rateLimit.middleware.js";
import { errorMiddleware } from "./middleware/error.middleware.js";
import morgan from "morgan";

import authRoutes from "./modules/auth/auth.routes.js";
import userRoutes from "./modules/users/user.routes.js";
import adminRoutes from "./modules/admin/admin.routes.js";
import projectRoutes from "./modules/projects/project.routes.js";
import eventsRoutes from "./modules/events/events.routes.js";
import contactRoutes from "./modules/contact/contact.routes.js";

import teamRoutes from "./modules/team/team.routes.js";
import alumniRoutes from "./modules/alumni/alumni.routes.js";
import blogsRoutes from "./modules/blogs/blogs.routes.js";
import achievementsRoutes from "./modules/achievements/achievements.routes.js";
import resourcesRoutes from "./modules/resources/resources.routes.js";

const app = express();

// ── Security & Parsing ────────────────────────────────────────────────────────
app.use(helmet());
app.use(cors({ origin: true, credentials: true }));
app.use(express.json({ limit: "50mb" }));
app.use(express.urlencoded({ extended: true, limit: "50mb" }));
app.use(morgan("dev"));

// ── Rate Limiting ─────────────────────────────────────────────────────────────
app.use("/api", apiLimiter);

// ── Swagger Docs ──────────────────────────────────────────────────────────────
app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerSpec));

// ── API Routes ────────────────────────────────────────────────────────────────
app.use("/api/v1/auth", authRoutes);
app.use("/api/v1/users", userRoutes);
app.use("/api/v1/admin", adminRoutes);
app.use("/api/v1/projects", projectRoutes);
app.use("/api/v1/events", eventsRoutes);
app.use("/api/v1/contact", contactRoutes);

app.use("/api/v1/team", teamRoutes);
app.use("/api/v1/alumni", alumniRoutes);
app.use("/api/v1/blogs", blogsRoutes);
app.use("/api/v1/achievements", achievementsRoutes);
app.use("/api/v1/resources", resourcesRoutes);

// ── Health Check ──────────────────────────────────────────────────────────────
app.get("/", (_req, res) => {
  res.status(200).json({ success: true, message: "Code Vimarsh API is running successfully on Vercel 🚀" });
});

app.get("/health", (_req, res) => {
  res
    .status(200)
    .json({ success: true, message: "Code Vimarsh API is healthy 🚀" });
});

// ── 404 Handler ───────────────────────────────────────────────────────────────
app.use((req, res) => {
  res.status(404).json({
    success: false,
    message: `Route ${req.originalUrl} not found.`,
  });
});

app.use(errorMiddleware);

export default app;